+++ cat_modules/db.js
@@ -10,14 +10,22 @@ module.exports = {
                name       TEXT,
                admin      BOOLEAN DEFAULT 0)`);
 
+      _db.run(`CREATE TABLE IF NOT EXISTS sellers (
+                id         INTEGER PRIMARY KEY,
+                name       TEXT,
+                location   TEXT,
+                userId     INTEGER,
+                FOREIGN KEY(userId)
+                REFERENCES users(id)
+                ON DELETE CASCADE)`);
+
       _db.run(`CREATE TABLE IF NOT EXISTS listings (
                id       INTEGER PRIMARY KEY,
                item     TEXT,
                price    TEXT,
-               userId   INTEGER,
-               location TEXT,
-               FOREIGN KEY(userId)
-               REFERENCES users(id)
+               sellerId INTEGER,
+               FOREIGN KEY(sellerId)
+               REFERENCES sellers(id)
                ON DELETE CASCADE)`);
     });
 
@@ -28,23 +36,3 @@ module.exports = {
     return _db;
   }
 }
-
-
-// _db.run(`CREATE TABLE IF NOT EXISTS sellers (
-//   id         INTEGER PRIMARY KEY,
-//   name       TEXT,
-//   location   TEXT,
-//   userId     INTEGER,
-//   FOREIGN KEY(userId)
-//   REFERENCES users(id)
-//   ON DELETE CASCADE)`);
-
-// _db.run(`CREATE TABLE IF NOT EXISTS listings (
-//   id       INTEGER PRIMARY KEY,
-//   item     TEXT,
-//   price    TEXT,
-//   sellerId INTEGER,
-//   FOREIGN KEY(sellerId)
-//   REFERENCES sellers(id)
-//   ON DELETE CASCADE)`);
-// });
