+++ cat_modules/parse_input.js
@@ -1,10 +1,7 @@
-exports.run = (message, action) => {
-  const pattern = {
-    "add": "(?<item>[^:]*):(?<price>[^:]*)",
-    "remove": "(?<listing>.*)",
-    "search": "(?<focus>(name|user|listings|seller)\\b)?\\s?(?<term>.*)"
-  }[action]
-
-  const args = message.content.match(`${action}\\s+${pattern}`);
+exports.run = (message) => {
+  // The responsibility of this is to take a command from a user
+  // and return the command, the sub-command, and the remainder as formatted args.
+  const args = message.content.match("\\s(?<command>[^\\s]*)\\s(?<subCommand>[^\\s]*)\\s?(?<args>.*)")
+  
   return args ? args.groups : {};
 }
