+++ commands/admin-listing-delete.js
